# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shop API",
    "version": "1.2.0",
    "description": "REST API for the Shop application"
  },
  "servers": [{ "url": "http://localhost:3001" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Review": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "userName": { "type": "string" },
          "productId": { "type": "integer" },
          "orderId": { "type": "integer" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Report": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "userId": { "type": "integer" },
          "reason": { "type": "string", "example": "not_received" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["open", "in_progress", "resolved", "rejected"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/reports": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get reports (admin: all, user: own)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of reports",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Report" } }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Reports"],
        "summary": "Create a report for a paid order (triggers auto-refund)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId", "reason"],
                "properties": {
                  "orderId": { "type": "integer", "example": 1 },
                  "reason": { "type": "string", "enum": ["not_received", "wrong_item", "damaged", "other"] },
                  "description": { "type": "string", "example": "Item never arrived" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created (refund auto-triggered if payment intent exists)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Report" }
              }
            }
          },
          "400": { "description": "Validation error or order not paid" },
          "403": { "description": "Access denied" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/reports/{id}": {
      "patch": {
        "tags": ["Reports"],
        "summary": "Update report status (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["open", "in_progress", "resolved", "rejected"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated report" },
          "400": { "description": "Invalid status" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Report not found" }
        }
      }
    },
    "/api/reports/{id}/refund": {
      "post": {
        "tags": ["Reports"],
        "summary": "Manually issue a Stripe refund for a report (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Refund issued" },
          "400": { "description": "Report already closed or order already refunded" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Report or order not found" },
          "500": { "description": "Stripe error" }
        }
      }
    },
    "/api/payments/create-session": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create a Stripe checkout session for an order",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId"],
                "properties": {
                  "orderId": { "type": "integer", "example": 1 },
                  "couponCode": { "type": "string", "example": "SAVE10" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns Stripe checkout URL" },
          "400": { "description": "Order already paid, expired, or invalid coupon" },
          "403": { "description": "Access denied" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/payments/invoice/{orderId}": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get Stripe invoice PDF URL for a paid order",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "orderId", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Returns invoice PDF URL" },
          "403": { "description": "Access denied" },
          "404": { "description": "Order or invoice not found" }
        }
      }
    },
    "/api/payments/webhook": {
      "post": {
        "tags": ["Payments"],
        "summary": "Stripe webhook endpoint (internal, not for direct use)",
        "responses": {
          "200": { "description": "Webhook received" },
          "400": { "description": "Invalid signature" }
        }
      }
    },
...
    "/api/products/{id}/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get reviews and average rating for a product",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Reviews list with averageRating and total",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reviews": { "type": "array", "items": { "$ref": "#/components/schemas/Review" } },
                    "averageRating": { "type": "number", "nullable": true },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": { "description": "Product not found" }
        }
      }
    },
...
    "/api/reviews/my": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get reviews written by the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of own reviews" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/reviews": {
      "post": {
        "tags": ["Reviews"],
        "summary": "Submit a review for a product",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["productId", "orderId", "rating"],
                "properties": {
                  "productId": { "type": "integer", "example": 1 },
                  "orderId": { "type": "integer", "example": 3 },
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5, "example": 4 },
                  "comment": { "type": "string", "example": "Great product!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Review created" },
          "400": { "description": "Order not paid or product not in order" },
          "401": { "description": "Unauthorized" },
          "409": { "description": "Already reviewed this product" }
        }
      }
    },
    "/api/reviews/{id}": {
      "put": {
        "tags": ["Reviews"],
        "summary": "Edit own review",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated review" },
          "403": { "description": "Access denied" },
          "404": { "description": "Review not found" }
        }
      },
      "delete": {
        "tags": ["Reviews"],
        "summary": "Delete own review (or any review if admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Review deleted" },
          "403": { "description": "Access denied" },
          "404": { "description": "Review not found" }
        }
      }
    }
  }
}